Refactor the Model tracking in a module
Contain the unsafe code in a smaller module.
Add an abstraction to make it easier to create change listener other
than the repeater.
